Renamed and organized the python modules. I also went into some of the jupyter files and renamed the imports
