+++ general_qec/errors.py
@@ -2,8 +2,8 @@
 
 import numpy as np
 import random
-from quantum.qec_helpers import *
-from quantum.gates import *
+from general_qec.qec_helpers import *
+from general_qec.gates import *
 
 ### Applies a random X rotation to one of the physical qubits in your system (randomly) ### 
 def random_qubit_x_error(logical_state, qubit_range = None):
@@ -18,15 +18,14 @@ def random_qubit_x_error(logical_state, qubit_range = None):
         if error_index != -1:
             error_index = random.randint(qubit_range[0], qubit_range[1]-1)
              
-    # Apply the errro depending on the index
+    # Apply the error depending on the index
     if error_index == -1:
         errored_logical_state = logical_state
-        return errored_logical_state, error_index
     else:   
         error_gate = np.kron(np.identity(2**(error_index)), np.kron(sigma_x, np.identity(2**(n-error_index-1))))
         errored_logical_state = np.dot(error_gate, logical_state)
 
-        return errored_logical_state, error_index
+    return errored_logical_state, error_index
 
 ### Applies a random Z rotation to one of the physical qubits in your system (randomly) ### 
 def random_qubit_z_error(logical_state, qubit_range = None):
@@ -41,15 +40,14 @@ def random_qubit_z_error(logical_state, qubit_range = None):
         if error_index != -1:
             error_index = random.randint(qubit_range[0], qubit_range[1]-1)
              
-    # Apply the errro depending on the index
+    # Apply the error depending on the index
     if error_index == -1:
         errored_logical_state = logical_state
-        return errored_logical_state, error_index
     else:   
         error_gate = np.kron(np.identity(2**(error_index)), np.kron(sigma_z, np.identity(2**(n-error_index-1))))
         errored_logical_state = np.dot(error_gate, logical_state)
 
-        return errored_logical_state, error_index
+    return errored_logical_state, error_index
 
 
 ### - - - - - - - - - - Gates which contain probability for errors (line connectivity) - - - - - - - - - - ###
